feat: machine Python package
