+++ machine/config.py
@@ -1,0 +1,7 @@
+"""Configure"""
+
+import logging
+import os
+
+if 'DEBUG' in os.environ:
+    logging.basicConfig(level=logging.DEBUG)
